+++ Lengine/InputManager.h
@@ -12,29 +12,29 @@ namespace Lengine {
 
 		void update();
 
-		void pressKey(unsigned int keyID);
-		void releaseKey(unsigned int keyID);
+		void pressKey(uint32_t keyID);
+		void releaseKey(uint32_t keyID);
 		void setMouseCoords(float x, float y);
-		bool isKeyPressed(unsigned int keyID);
-		bool isKeyDown(unsigned int keyID);
+		bool isKeyPressed(uint32_t keyID);
+		bool isKeyDown(uint32_t keyID);
 		
-		void pressMouseButton(unsigned int buttonID);
-		void releaseMouseButton(unsigned int buttonID);
-		bool isMouseButtonDown(unsigned int buttonID);
+		void pressMouseButton(uint8_t buttonID);
+		void releaseMouseButton(uint8_t buttonID);
+		bool isMouseButtonDown(uint8_t buttonID);
 
-		bool isMouseButtonPressed(unsigned int buttonID);
-		bool wasMouseButtonDown(unsigned int buttonID);
+		bool isMouseButtonPressed(uint8_t buttonID);
+		bool wasMouseButtonDown(uint8_t buttonID);
 
 		glm::vec2 getMouseCoords() const { return _mouseCoords; }
 
 
 	private:
-		bool wasKeyDown(unsigned int keyID);
-		std::unordered_map<unsigned int, bool> _keyMap;
-		std::unordered_map<unsigned int, bool> _previousKeyMap;
+		bool wasKeyDown(uint32_t keyID);
+		std::unordered_map<uint32_t, bool> _keyMap;
+		std::unordered_map<uint32_t, bool> _previousKeyMap;
 
-		std::unordered_map<unsigned int, bool> _mouseMap;
-		std::unordered_map<unsigned int, bool> _previousMouseMap;
+		std::unordered_map<uint8_t, bool> _mouseMap;
+		std::unordered_map<uint8_t, bool> _previousMouseMap;
 
 		glm::vec2 _mouseCoords;
